\name{blood.pressure}
\alias{blood.pressure}
\docType{data}
\title{Blood Pressure Data for qrmix
}
\description{
Simulated blood pressure data created for usage in qrmix examples.
}
\usage{blood.pressure}
\format{
  A data frame with 500 observations on the following 7 variables.
  \describe{
    \item{\code{bmi}}{a numeric vector referring to body mass index}
    \item{\code{age}}{a numeric vector}
    \item{\code{systolic}}{a numeric vector referring to systolic blood pressure}
    \item{\code{diastolic}}{a numeric vector referring to diastolic blood pressure}
    \item{\code{gender}}{a factor with levels \code{female} and \code{male}}
    \item{\code{race}}{a factor with levels \code{white}, \code{black}, and \code{other}}
    \item{\code{smoking}}{a factor with levels \code{yes} and \code{no}}
  }
}

\note{
This data does not include any real patient information.
}

\keyword{datasets}
